requirements.txt
astroid==2.3.3
Babel==2.8.0
click==7.1.1
colorama==0.4.3
dominate==2.5.1
Flask==1.1.1
Flask-Babel==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Table==0.5.0
Flask-WTF==0.14.3
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.4
pylint==2.4.4
python-dotenv==0.12.0
pytz==2019.3
six==1.14.0
typed-ast==1.4.1
visitor==0.1.3
Werkzeug==1.0.0
wrapt==1.11.2
WTForms==2.2.1
WTForms-Components==0.10.4
loguru==0.4.1